requirements.txt
urllib3
requests
bs4
lxml
openpyxl
selenium
fake-useragent
python-dotenv
pyinstaller